maps_search_agent/requirements.txt
pyautogui==0.9.54
pillow==10.0.1
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
requests==2.31.0
pyperclip==1.8.2
keyboard==0.13.5
mouse==0.7.1
psutil==5.9.5
screeninfo==0.8.1
imageio==2.31.5
scikit-image==0.21.0
scipy==1.11.4
